App/models/db.py
import mysql.connector

def get_connection():
    mydb = mysql.connector.connect(
        host="localhost",
        user="Admin",
        password="shaaragy970",
        database="smca"
    )
    return mydb