requirements.txt
anyio==3.7.1
blinker==1.6.2
certifi==2023.5.7
click==8.1.4
colorama==0.4.6
dnspython==2.4.0
email-validator==2.0.0.post2
Flask==2.3.2
Flask-WTF==1.1.1
h11==0.14.0
httpcore==0.17.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.0.33
protobuf==3.20.3
sniffio==1.3.0
Werkzeug==2.3.6
WTForms==3.0.1